user-service/requirements.txt
flask==2.3.2
pyjwt==2.8.0
psycopg2-binary==2.9.9
